server: refactor server code
