+++ server/BUILD.bazel
@@ -144,6 +144,7 @@ go_test(
         "extract_test.go",
         "main_test.go",
         "mock_conn_test.go",
+        "optimize_trace_test.go",
         "stat_test.go",
         "tidb_library_test.go",
         "tidb_serial_test.go",
@@ -192,6 +193,7 @@ go_test(
         "//util/cpuprofile",
         "//util/dbterror/exeerrors",
         "//util/plancodec",
+        "//util/replayer",
         "//util/resourcegrouptag",
         "//util/stmtsummary/v2:stmtsummary",
         "//util/syncutil",
