+++ server/planreplayer/BUILD.bazel
@@ -35,24 +35,17 @@ go_test(
     ],
     embed = [":planreplayer"],
     flaky = True,
-    shard_count = 3,
     deps = [
         "//domain",
         "//server",
-        "//server/internal",
         "//server/internal/testutil",
         "//server/internal/util",
         "//server/statshandler",
         "//session",
         "//statistics/handle",
         "//testkit",
-        "//testkit/testdata",
-        "//util/arena",
-        "//util/chunk",
-        "//util/replayer",
         "@com_github_go_sql_driver_mysql//:mysql",
         "@com_github_gorilla_mux//:mux",
-        "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
         "@org_golang_x_exp//slices",
     ],
