+++ server/planreplayer/optimize_trace_test.go
@@ -15,28 +15,18 @@
 package planreplayer
 
 import (
-	"bufio"
-	"bytes"
-	"context"
 	"database/sql"
-	"encoding/json"
 	"net/http"
-	"os"
 	"path/filepath"
 	"testing"
 
 	"github.com/gorilla/mux"
-	"github.com/pingcap/failpoint"
 	server2 "github.com/pingcap/tidb/server"
-	"github.com/pingcap/tidb/server/internal"
+	"github.com/pingcap/tidb/server/internal/testutil"
 	"github.com/pingcap/tidb/server/internal/util"
 	"github.com/pingcap/tidb/server/statshandler"
 	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/testkit"
-	"github.com/pingcap/tidb/testkit/testdata"
-	"github.com/pingcap/tidb/util/arena"
-	"github.com/pingcap/tidb/util/chunk"
-	"github.com/pingcap/tidb/util/replayer"
 	"github.com/stretchr/testify/require"
 )
 
@@ -44,10 +34,10 @@ func TestDumpOptimizeTraceAPI(t *testing.T) {
 	store := testkit.CreateMockStore(t)
 
 	driver := server2.NewTiDBDriver(store)
-	client := newTestServerClient()
+	client := testutil.NewTestServerClient()
 	cfg := util.NewTestConfig()
-	cfg.Port = client.port
-	cfg.Status.StatusPort = client.statusPort
+	cfg.Port = client.GetPort()
+	cfg.Status.StatusPort = client.GetStatusPort()
 	cfg.Status.ReportStatus = true
 
 	server, err := server2.NewServer(cfg, driver)
@@ -58,15 +48,15 @@ func TestDumpOptimizeTraceAPI(t *testing.T) {
 	require.NoError(t, err)
 	server.SetDomain(dom)
 
-	client.port = getPortFromTCPAddr(server.listener.Addr())
-	client.statusPort = getPortFromTCPAddr(server.statusListener.Addr())
+	client.SetPort(testutil.GetPortFromTCPAddr(server.ListenerAddr()))
+	client.SetStatusPort(testutil.GetPortFromTCPAddr(server.StatusListenerAddr()))
 	go func() {
 		err := server.Run()
 		require.NoError(t, err)
 	}()
-	client.waitUntilServerOnline()
+	client.WaitUntilServerOnline()
 
-	statsHandler := &statshandler.StatsHandler{dom}
+	statsHandler := statshandler.NewStatsHandler(dom)
 
 	otHandler := &OptimizeTraceHandler{}
 	filename := prepareData4OptimizeTrace(t, client, statsHandler)
@@ -74,7 +64,7 @@ func TestDumpOptimizeTraceAPI(t *testing.T) {
 	router := mux.NewRouter()
 	router.Handle("/optimize_trace/dump/{filename}", otHandler)
 
-	resp0, err := client.fetchStatus(filepath.Join("/optimize_trace/dump/", filename))
+	resp0, err := client.FetchStatus(filepath.Join("/optimize_trace/dump/", filename))
 	require.NoError(t, err)
 	defer func() {
 		require.NoError(t, resp0.Body.Close())
@@ -82,8 +72,8 @@ func TestDumpOptimizeTraceAPI(t *testing.T) {
 	require.Equal(t, http.StatusOK, resp0.StatusCode)
 }
 
-func prepareData4OptimizeTrace(t *testing.T, client *testServerClient, statHandle *statshandler.StatsHandler) string {
-	db, err := sql.Open("mysql", client.getDSN())
+func prepareData4OptimizeTrace(t *testing.T, client *testutil.TestServerClient, statHandle *statshandler.StatsHandler) string {
+	db, err := sql.Open("mysql", client.GetDSN())
 	require.NoError(t, err, "Error connecting")
 	defer func() {
 		err := db.Close()
@@ -91,7 +81,7 @@ func prepareData4OptimizeTrace(t *testing.T, client *testServerClient, statHandl
 	}()
 	tk := testkit.NewDBTestKit(t, db)
 
-	h := statHandle.do.StatsHandle()
+	h := statHandle.Domain().StatsHandle()
 	tk.MustExec("create database optimizeTrace")
 	tk.MustExec("use optimizeTrace")
 	tk.MustExec("create table t(a int)")
@@ -103,71 +93,4 @@ func prepareData4OptimizeTrace(t *testing.T, client *testServerClient, statHandl
 	err = rows.Scan(&filename)
 	require.NoError(t, err)
 	return filename
-}
-
-func TestOptimizerDebugTrace(t *testing.T) {
-	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/planner/SetBindingTimeToZero", `return(true)`))
-	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/planner/core/DebugTraceStableStatsTbl", `return(true)`))
-	defer func() {
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/planner/SetBindingTimeToZero"))
-		require.NoError(t, failpoint.Disable("github.com/pingcap/tidb/planner/core/DebugTraceStableStatsTbl"))
-	}()
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-	tk := testkit.NewTestKit(t, store)
-	tidbdrv := server2.NewTiDBDriver(store)
-	cfg := util.NewTestConfig()
-	cfg.Port, cfg.Status.StatusPort = 0, 0
-	cfg.Status.ReportStatus = false
-	server, err := server2.NewServer(cfg, tidbdrv)
-	require.NoError(t, err)
-	defer server.Close()
-	cc := &server2.clientConn{
-		server:     server,
-		alloc:      arena.NewAllocator(1024),
-		chunkAlloc: chunk.NewAllocator(),
-		pkt:        internal.NewPacketIOForTest(bufio.NewWriter(bytes.NewBuffer(nil))),
-	}
-	ctx := context.Background()
-	cc.setCtx(&server2.TiDBContext{Session: tk.Session(), stmts: make(map[int]*server2.TiDBStatement)})
-
-	tk.MustExec("use test")
-	tk.MustExec("create table t (col1 int, index i(col1))")
-	h := dom.StatsHandle()
-	require.NoError(t, h.HandleDDLEvent(<-h.DDLEventCh()))
-
-	tk.MustExec("plan replayer capture '0595c79f25d183319d0830ff8ca538c9054cbf407e5e27488b5dc40e4738a7c8' '*'")
-	tk.MustExec("plan replayer capture 'c0fcc0abbaaffcaafe21115a3c67ae5d96a188cc197559953d2865ea6852d3cc' '*'")
-	tk.MustExec("plan replayer capture '58fcbdd56a722c02225488c89a782cd2d626f8219c8ef8f57cd3bcdb6eb7c1b2' '*'")
-	require.NoError(t, cc.handleStmtPrepare(ctx, "select sum(col1) from t where col1 < ? and col1 > 100"))
-	tk.MustExec("prepare stmt from 'select * from t where col1 in (?, 2, 3)'")
-	tk.MustExec("set @a = 1")
-	var (
-		in  []string
-		out []interface{}
-	)
-	optSuiteData := server2.testDataMap["optimizer_suite"]
-	optSuiteData.LoadTestCases(t, &in, &out)
-	for i, cmdString := range in {
-		require.NoError(t, cc.dispatch(ctx, []byte(cmdString)))
-		traceInfo := cc.ctx.Session.GetSessionVars().StmtCtx.OptimizerDebugTrace
-		var buf bytes.Buffer
-		encoder := json.NewEncoder(&buf)
-		encoder.SetEscapeHTML(false)
-		require.NoError(t, encoder.Encode(traceInfo))
-		var res interface{}
-		require.NoError(t, json.Unmarshal(buf.Bytes(), &res))
-		testdata.OnRecord(func() {
-			out[i] = res
-		})
-		require.Equal(t, out[i], res, cmdString)
-	}
-
-	prHandle := dom.GetPlanReplayerHandle()
-	worker := prHandle.GetWorker()
-	for i := 0; i < 3; i++ {
-		task := prHandle.DrainTask()
-		success := worker.HandleTask(task)
-		require.True(t, success)
-		require.NoError(t, os.Remove(filepath.Join(replayer.GetPlanReplayerDirName(), task.FileName)))
-	}
 }
